requirements.txt
Flask
flask-wtf
sqlalchemy
pymysql
twilio
pycryptodome
pycryptodomex
gunicorn